+++ include/open_atmos/types.hpp
@@ -8,6 +8,7 @@
 #include <optional>
 #include <string>
 #include <unordered_map>
+#include <map>
 #include <vector>
 
 namespace open_atmos
